+++ include/common.h
@@ -18,6 +18,8 @@
 
      https://www.apache.org/licenses/LICENSE-2.0
 
+   SPDX-License-Identifier: Apache-2.0
+
    Gather some functions common to multiple executables
 
    - detect_file_args