+++ include/xxhash.h
@@ -1,3 +1,4 @@
+/* SPDX-License-Identifier: BSD-2-Clause */
 /*
  * xxHash - Extremely Fast Hash algorithm
  * Header File