+++ instrumentation/afl-gcc-cmptrs-pass.so.cc
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
 /* GCC plugin for cmplog routines instrumentation of code for AFL++.
 
    Copyright 2014-2019 Free Software Foundation, Inc