+++ instrumentation/afl-llvm-lto-instrumentlist.so.cc
@@ -14,6 +14,8 @@
 
      https://www.apache.org/licenses/LICENSE-2.0
 
+   SPDX-License-Identifier: Apache-2.0
+
  */
 
 #define AFL_LLVM_PASS