+++ src/afl-cc.c
@@ -13,6 +13,8 @@
 
      https://www.apache.org/licenses/LICENSE-2.0
 
+   SPDX-License-Identifier: Apache-2.0
+
  */
 
 #define AFL_MAIN
@@ -3852,6 +3854,12 @@ static void edit_params(aflcc_state_t *aflcc, u32 argc, char **argv,
       }
 
     }
+
+// link in execinfo on FreeBSD to include backtrace library used by
+// instrumentation.
+#ifdef __FreeBSD__
+    insert_param(aflcc, "-lexecinfo");
+#endif
 
   }