+++ src/afl-cmin.c
@@ -20,6 +20,8 @@
 
      https://www.apache.org/licenses/LICENSE-2.0
 
+   SPDX-License-Identifier: Apache-2.0
+
    A tool to minimize the corpus.
 
  */