+++ src/afl-common.c
@@ -17,6 +17,8 @@
 
      https://www.apache.org/licenses/LICENSE-2.0
 
+   SPDX-License-Identifier: Apache-2.0
+
    Gather some functions common to multiple executables
 
    - detect_file_args